+++ tomcat/src/main/java/org/apache/coyote/http11/request/RequestHandler.java
@@ -1,8 +1,8 @@
 package org.apache.coyote.http11.request;
 
-import org.apache.coyote.http11.auth.AuthService;
+import org.apache.coyote.http11.auth.LoginController;
+import org.apache.coyote.http11.auth.RegisterController;
 import org.apache.coyote.http11.request.line.Protocol;
-import org.apache.coyote.http11.request.line.RequestLine;
 import org.apache.coyote.http11.response.Location;
 import org.apache.coyote.http11.response.ResponseEntity;
 
@@ -10,21 +10,18 @@ import static org.apache.coyote.http11.response.HttpStatus.OK;
 
 public class RequestHandler {
 
-    private final AuthService authService = new AuthService();
+    private final RegisterController registerController = new RegisterController();
+    private final LoginController loginController = new LoginController();
 
     public ResponseEntity getResponse(Request request) {
-        final RequestLine requestLine = request.requestLine();
-        final RequestHeader requestHeader = request.requestHeader();
-        final RequestBody requestBody = request.requestBody();
-
         final Protocol protocol = request.requestLine().protocol();
         final String path = request.requestLine().path().defaultPath();
 
-        if (path.equals("/login")) {
-            return authService.login(requestLine, requestHeader, requestBody);
-        }
         if (path.equals("/register")) {
-            return authService.register(requestLine, requestBody);
+            return registerController.register(request);
+        }
+        if (path.equals("/login")) {
+            return loginController.login(request);
         }
         return ResponseEntity.getCookieNullResponseEntity(protocol, OK, Location.from(path));
     }
